requirements.txt
requests
python-whois
dnspython
urllib3